refactor(native): extract certificate material I/O into zkrio
Modified AI generated code: YES
Dev Name: Fernando Rijo Cedeno, Reviewed by: TBD
