+++ native/c/commands/certificates.cpp
@@ -11,8 +11,8 @@
 
 #include "certificates.hpp"
 #include "../zkr.hpp"
+#include "../zkrio.hpp"
 #include "../zbase64.h"
-#include "../zds.hpp"
 #include "../ztype.h"
 #include "../zut.hpp"
 #include <cctype>
@@ -107,91 +107,6 @@ int run_simple(InvocationContext &context, int rc, ZKR &zkr, const std::string &
     result->set("warning", str(zkr.diag.warning));
   add_saf_returns(result, zkr);
   context.set_object(result);
-  return RTNCD_SUCCESS;
-}
-
-// Read a PKCS#12 blob out of a sequential data set or PDS/E member. Binary mode so
-// LE concatenates V-format record data with no RDWs and no code-page conversion --
-// byte-identical to `cp -B "//'DSN'"`.
-int read_cert_dsn(const std::string &dsn, std::string &data, std::string &err)
-{
-  if (!zds_dataset_exists(dsn))
-  {
-    err = "Could not access data set: " + dsn;
-    return RTNCD_FAILURE;
-  }
-
-  ZDS zds{};
-  zut_prepare_encoding("binary", &zds.encoding_opts);
-  const int rc = zds_read(ZDSReadOpts{.zds = &zds, .dsname = dsn}, data);
-  if (rc != RTNCD_SUCCESS)
-  {
-    err = zds.diag.e_msg;
-    return RTNCD_FAILURE;
-  }
-  if (data.empty())
-  {
-    err = "PKCS#12 data set is empty: " + dsn;
-    return RTNCD_FAILURE;
-  }
-  return RTNCD_SUCCESS;
-}
-
-// Write exported certificate bytes to a data set, creating it when absent.
-// is_binary=true  (p12): raw bytes, chunked into V-format records (zds_write_binary).
-// is_binary=false (pem): the EBCDIC PEM text is written as records via zds_write in
-//                        text mode -- one line per record.
-int write_cert_dsn(const std::string &dsn, const std::string &data, bool is_binary, std::string &err)
-{
-  if (!zds_dataset_exists(dsn))
-  {
-    const auto open_paren = dsn.find('(');
-    const auto close_paren = dsn.find(')');
-    const bool has_member = open_paren != std::string::npos && close_paren != std::string::npos && close_paren > open_paren;
-    const std::string base_dsn = has_member ? dsn.substr(0, open_paren) : dsn;
-
-    if (has_member)
-    {
-      const std::string member = dsn.substr(open_paren + 1, close_paren - open_paren - 1);
-      if (!zds_is_valid_member_name(member))
-      {
-        err = "Invalid member name: " + member;
-        return RTNCD_FAILURE;
-      }
-    }
-
-    // RACDCERT FORMAT(PKCS12DER) parity: PS/VB/LRECL(84)/BLKSIZE(27998).
-    DS_ATTRIBUTES a{}; // zero-init: zds_create_dsn preserves 0, only negative means "unset"
-    a.dsorg = has_member ? "PO" : "PS";
-    a.recfm = "V,B"; // mandatory -- FB pads records with blanks, which would change the byte count
-    a.lrecl = 84;
-    a.blksize = 27998;
-    a.alcunit = "TRK";
-    a.primary = 5;
-    a.secondary = 5;
-    if (has_member)
-    {
-      a.dirblk = 5;
-      a.dsntype = ZDS_DSNTYPE_LIBRARY;
-    }
-
-    std::string response;
-    if (zds_create_dsn(nullptr, base_dsn, a, response) != RTNCD_SUCCESS)
-    {
-      err = "Could not create data set '" + base_dsn + "': " + response;
-      return RTNCD_FAILURE;
-    }
-  }
-
-  ZDS zds{}; // eDataTypeText + empty codepage -> zds_use_codepage() is false -> no iconv for PEM
-  const int rc = is_binary
-                     ? zds_write_binary(ZDSWriteOpts{.zds = &zds, .dsname = dsn}, data)
-                     : zds_write(ZDSWriteOpts{.zds = &zds, .dsname = dsn}, data);
-  if (rc != RTNCD_SUCCESS)
-  {
-    err = zds.diag.e_msg;
-    return RTNCD_FAILURE;
-  }
   return RTNCD_SUCCESS;
 }
 
@@ -338,7 +253,7 @@ int handle_cert_export(InvocationContext &context)
     // the file is created (or truncated) owner-read/write only, independent of
     // the process umask.
     std::string err;
-    if (zut_write_file_private(file, data, err) != RTNCD_SUCCESS)
+    if (zkrio_write_file(file, data, err) != RTNCD_SUCCESS)
     {
       context.error_stream() << "Error: could not write output file: " << file << " (" << err << ")" << std::endl;
       return RTNCD_FAILURE;
@@ -351,7 +266,7 @@ int handle_cert_export(InvocationContext &context)
   else if (!dsn.empty())
   {
     std::string err;
-    if (write_cert_dsn(dsn, data, is_p12, err) != RTNCD_SUCCESS)
+    if (zkrio_write_dsn(dsn, data, is_p12, err) != RTNCD_SUCCESS)
     {
       context.error_stream() << "Error: could not write output data set: " << dsn << " (" << err << ")" << std::endl;
       return RTNCD_FAILURE;
@@ -428,7 +343,7 @@ int handle_cert_import(InvocationContext &context)
   if (!dsn.empty())
   {
     std::string err;
-    if (read_cert_dsn(dsn, opts.p12_data, err) != RTNCD_SUCCESS)
+    if (zkrio_read_dsn(dsn, opts.p12_data, err) != RTNCD_SUCCESS)
     {
       context.error_stream() << "Error: could not read source data set: " << dsn << " (" << err << ")" << std::endl;
       return RTNCD_FAILURE;
